pyproject.toml
[project]
name = "drama-web-crawler"
version = "0.1.0"
description = "Daily drama crawler with CSV exports and FastAPI download endpoints."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.4",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
